Refactor to keep tests in the right __tests__ folders
