+++ src/test-utilities/WithTheme/index.tsx
@@ -1,5 +1,5 @@
 import { ThemeProvider } from "styled-components";
-import { modalTheme } from "../../../components/Theme";
+import { modalTheme } from "../../components/Theme";
 import { WithThemeProps } from "./types";
 
 const WithTheme = ({ children }: WithThemeProps) => (
